fix: steam library games are not the steam client
exes under steamapps/common classified as the steam client and got its
compat profile. detection moves into WhiskyKit (LauncherType.detect),
the app delegates to it, and the test file stops testing a hand-copied
duplicate that could never fail against the real code. program-list and
pin launches now run detection like the other launch paths.

part of #158
