+++ Whisky/Utils/LauncherDetection.swift
@@ -103,79 +103,7 @@ enum LauncherDetection {
     /// - Parameter url: The URL to the Windows executable file
     /// - Returns: The detected launcher type, or `nil` if no launcher detected
     static func detectLauncher(from url: URL) -> LauncherType? {
-        let filename = url.lastPathComponent.lowercased()
-        let path = url.path.lowercased()
-
-        // Steam detection
-        // Common paths: steam.exe, steamservice.exe, steamwebhelper.exe
-        if filename.contains("steam") || path.contains("/steam/") || path.contains("\\steam\\") {
-            return .steam
-        }
-
-        // Rockstar Games Launcher detection
-        // Common paths: Launcher.exe in Rockstar Games directory
-        // LauncherPatcher.exe (workaround for whisky-app/whisky#835)
-        // Note: Be specific about generic "launcher.exe" to avoid false positives
-        if filename.contains("rockstar") ||
-            filename.contains("launcherpatcher") ||
-            path.contains("rockstar games") ||
-            path.contains("rockstar games launcher") ||
-            (filename == "launcher.exe" &&
-                (path.contains("rockstar games") || path.contains("social club"))) {
-            return .rockstar
-        }
-
-        // EA App / Origin detection
-        // EADesktop.exe (new EA App), Origin.exe (legacy)
-        if filename.contains("eadesktop") ||
-            filename.contains("eaapp") ||
-            filename.contains("origin.exe") ||
-            path.contains("/ea app/") ||
-            path.contains("\\ea app\\") ||
-            path.contains("/origin/") {
-            return .eaApp
-        }
-
-        // Epic Games Store detection
-        // EpicGamesLauncher.exe, EpicWebHelper.exe
-        if filename.contains("epicgames") ||
-            filename.contains("epiclauncher") ||
-            filename.contains("epicwebhelper") ||
-            path.contains("/epic games/") ||
-            path.contains("\\epic games\\") {
-            return .epicGames
-        }
-
-        // Ubisoft Connect detection
-        // UbisoftConnect.exe, upc.exe (legacy Uplay)
-        if filename.contains("ubisoft") ||
-            filename.contains("uplay") ||
-            filename.contains("upc.exe") ||
-            path.contains("/ubisoft") {
-            return .ubisoft
-        }
-
-        // Battle.net detection
-        // Battle.net.exe, Battle.net Launcher.exe
-        if filename.contains("battle.net") ||
-            filename.contains("battlenet") ||
-            path.contains("/battle.net/") ||
-            path.contains("\\battle.net\\") {
-            return .battleNet
-        }
-
-        // Paradox Launcher detection
-        // Paradox Launcher.exe - Be specific to avoid false positives
-        // Common path: C:/Users/User/AppData/Local/Programs/Paradox Launcher/
-        if filename.contains("paradox launcher") ||
-            filename.contains("paradoxlauncher") ||
-            path.contains("paradox launcher") ||
-            ((filename == "launcher.exe" || filename == "launcher") &&
-                path.contains("paradox interactive")) {
-            return .paradox
-        }
-
-        return nil
+        LauncherType.detect(from: url)
     }
 
     /// Applies launcher-specific fixes when running a program.
