+++ Whisky/Views/Programs/ProgramMenuView.swift
@@ -27,6 +27,9 @@ struct ProgramMenuView: View {
     var body: some View {
         Button("button.run", systemImage: "play") {
             Telemetry.capture(.firstProgramLaunchAttempted)
+            // Program-list and pin launches historically skipped launcher
+            // detection; only FileOpenView and the bottle Run button had it.
+            LauncherDetection.detectAndApplyLauncherFixes(from: program.url, for: program.bottle)
             program.run()
         }
         .labelStyle(.titleAndIcon)
